context: extract `git_repo.rs`
