+++ src/context/mod.rs
@@ -9,7 +9,6 @@ use clap::Parser;
 use gix::{
     Repository, ThreadSafeRepository,
     sec::{self as git_sec, trust::DefaultForLevel},
-    state as git_state,
 };
 #[cfg(test)]
 use std::collections::HashMap;
@@ -32,6 +31,10 @@ pub use crate::utils::env::Env;
 pub use crate::utils::statusline::{
     ClaudeCodeData, ContextWindow, CostInfo, CurrentUsage, ModelInfo, Workspace,
 };
+
+mod git_repo;
+
+pub use git_repo::{GitRemote, GitRepo};
 
 /// Context contains data or common methods that may be used by multiple modules.
 /// The data contained within Context will be relevant to this particular rendering
@@ -714,86 +717,6 @@ impl DirContents {
     }
 }
 
-pub struct GitRepo {
-    pub repo: ThreadSafeRepository,
-
-    /// If `current_dir` is a git repository or is contained within one,
-    /// this is the short name of the current branch name of that repo,
-    /// i.e. `main`.
-    pub branch: Option<String>,
-
-    /// If `current_dir` is a git repository or is contained within one,
-    /// this is the path to the root of that repo.
-    pub workdir: Option<PathBuf>,
-
-    /// The path of the repository's `.git` directory.
-    pub path: PathBuf,
-
-    /// State
-    pub state: Option<git_state::InProgress>,
-
-    /// Remote repository
-    pub remote: Option<Remote>,
-
-    /// Contains `true` if the value of `core.fsmonitor` is set to `true`.
-    /// If not `true`, `fsmonitor` is explicitly disabled in git commands.
-    pub(crate) fs_monitor_value_is_true: bool,
-}
-
-impl GitRepo {
-    /// Opens the associated git repository.
-    pub fn open(&self) -> Repository {
-        self.repo.to_thread_local()
-    }
-
-    /// Wrapper to execute external git commands.
-    /// Handles adding the appropriate `--git-dir` and `--work-tree` flags to the command.
-    /// Also handles additional features required for security, such as disabling `fsmonitor`.
-    /// At this time, mocking is not supported.
-    pub fn exec_git<T: AsRef<OsStr> + Debug>(
-        &self,
-        context: &Context,
-        git_args: impl IntoIterator<Item = T>,
-    ) -> Option<CommandOutput> {
-        let mut command = create_command("git").ok()?;
-
-        // A value of `true` should not execute external commands.
-        let fsm_config_value = if self.fs_monitor_value_is_true {
-            "core.fsmonitor=true"
-        } else {
-            "core.fsmonitor="
-        };
-
-        command.env("GIT_OPTIONAL_LOCKS", "0").args([
-            OsStr::new("-C"),
-            context.current_dir.as_os_str(),
-            OsStr::new("--git-dir"),
-            self.path.as_os_str(),
-            OsStr::new("-c"),
-            OsStr::new(fsm_config_value),
-        ]);
-
-        // Bare repositories might not have a workdir, so we need to check for that.
-        if let Some(wt) = self.workdir.as_ref() {
-            command.args([OsStr::new("--work-tree"), wt.as_os_str()]);
-        }
-
-        command.args(git_args);
-        log::trace!("Executing git command: {command:?}");
-
-        exec_timeout(
-            &mut command,
-            Duration::from_millis(context.root_config.command_timeout),
-        )
-    }
-}
-
-/// Remote repository
-pub struct Remote {
-    pub branch: Option<String>,
-    pub name: Option<String>,
-}
-
 // A struct of Criteria which will be used to verify current PathBuf is
 // of X language, criteria can be set via the builder pattern
 pub struct ScanDir<'a> {
@@ -921,7 +844,7 @@ fn get_current_branch(repository: &Repository) -> Option<gix::refs::FullName> {
 fn get_remote_repository_info(
     repository: &Repository,
     branch_name: Option<&gix::refs::FullNameRef>,
-) -> Option<Remote> {
+) -> Option<GitRemote> {
     let branch_name = branch_name?;
     let branch = repository
         .branch_remote_ref_name(branch_name, gix::remote::Direction::Fetch)
@@ -931,7 +854,7 @@ fn get_remote_repository_info(
         .branch_remote_name(branch_name.shorten(), gix::remote::Direction::Fetch)
         .map(|n| n.as_bstr().to_string());
 
-    Some(Remote { branch, name })
+    Some(GitRemote { branch, name })
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
